requirements.txt
openai>=1.0.0
python-dotenv
pyyaml
pydantic